Create write a code of c++ for inserting deleting and replacing inserted values
This code allows the user to insert, delete and replace values in an array. It first prompts the user to enter the size of the array and the values in it. Then it asks the user to provide the position and value to insert into the array. If the position is valid, it inserts the value into the array and updates the size of the array. It then prompts the user to enter the position to delete a value. If the position is valid, it deletes the value at that position and updates the size of the array. Finally, it prompts the user to enter the position and value to replace in the array. If the position is valid, it replaces the value at that position with the new value