Create C++ code to demonstrate QuickSort using a stack
This code uses the QuickSort algorithm to sort an array of integers. The swap function swaps two elements in the array, and the partition function partitions the array into two sub-arrays based on a pivot element. The quickSort function uses a stack to keep track of the sub-arrays that need to be partitioned and sorted. It pushes the low and high indices of the initial sub-array onto the stack, and then iteratively partitions and sorts sub-arrays until the stack is empty.